Add Input component into primary-beacon-use page
